+++ app/src/main/AndroidManifest.xml
@@ -8,32 +8,32 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:theme="@style/FriskyAppTheme">
-        <activity android:name=".ui.orders.OrderActivity"></activity>
-        <activity android:name=".ui.orders.MenuActivity" />
-        <activity android:name=".ui.profile.SetupProfileActivity" />
-        <activity android:name=".ui.SettingsActivity" />
-        <activity android:name=".ui.LoginActivity">
+        <activity android:name=".orders.OrderActivity"></activity>
+        <activity android:name=".orders.MenuActivity" />
+        <activity android:name=".profile.SetupProfileActivity" />
+        <activity android:name=".SettingsActivity" />
+        <activity android:name=".LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ui.SignUpActivity" />
+        <activity android:name=".SignUpActivity" />
         <activity
-            android:name=".ui.HomeActivity"
+            android:name=".HomeActivity"
             android:windowSoftInputMode="adjustPan">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
             </intent-filter>
         </activity>
-        <activity android:name=".ui.pubs.PubActivity" />
-        <activity android:name=".ui.profile.ProfileActivity" />
+        <activity android:name=".pubs.PubActivity" />
+        <activity android:name=".profile.ProfileActivity" />
         <activity
-            android:name=".ui.profile.EditProfileActivity"
+            android:name=".profile.EditProfileActivity"
             android:windowSoftInputMode="adjustPan" />
-        <activity android:name=".ui.profile.ViewUserActivity" />
-        <activity android:name=".ui.social.ChatActivity" />
+        <activity android:name=".profile.ViewUserActivity" />
+        <activity android:name=".social.ChatActivity" />
     </application>
 
 </manifest>
