+++ app/src/main/java/com/frisky/icebreaker/profile/EditProfileActivity.java
@@ -1,9 +1,9 @@
-package com.frisky.icebreaker.ui.social;
+package com.frisky.icebreaker.profile;
 
 import android.graphics.Typeface;
+import android.os.Bundle;
 import android.support.v4.content.res.ResourcesCompat;
 import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
 import android.util.TypedValue;
 import android.view.View;
 import android.widget.ImageButton;
@@ -11,26 +11,23 @@ import android.widget.TextView;
 
 import com.frisky.icebreaker.R;
 
-public class ChatActivity extends AppCompatActivity {
+public class EditProfileActivity extends AppCompatActivity {
 
-    private TextView mToolbarText;
-    private ImageButton mBackButton;
-    private ImageButton mMenuButton;
+    TextView mToolbarText;
+    ImageButton mBackButton;
+    ImageButton mDoneButton;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_chat);
+        setContentView(R.layout.activity_edit_profile);
 
         initUI();
     }
 
     private void initUI() {
         mToolbarText = findViewById(R.id.text_app_bar);
-        if (getIntent().hasExtra("name")){
-            String name = getIntent().getStringExtra("name");
-            mToolbarText.setText(name);
-        }
+        mToolbarText.setText(R.string.edit_profile);
         mToolbarText.setTextSize(TypedValue.COMPLEX_UNIT_SP, 20);
         Typeface typeface = ResourcesCompat.getFont(getApplicationContext(), R.font.museosans700);
         mToolbarText.setTypeface(typeface);
@@ -40,11 +37,16 @@ public class ChatActivity extends AppCompatActivity {
         mBackButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                ChatActivity.super.onBackPressed();
+                EditProfileActivity.super.onBackPressed();
             }
         });
 
-        mMenuButton = findViewById(R.id.button_app_bar_right);
-        mMenuButton.setImageResource(R.drawable.round_more_vert_24);
+        getSupportFragmentManager()
+                .beginTransaction()
+                .replace(R.id.frame_image_grid, new EditImagesFragment())
+                .commit();
+
+        mDoneButton = findViewById(R.id.button_app_bar_right);
+        mDoneButton.setImageResource(R.drawable.round_done_24);
     }
 }
