+++ app/src/main/java/com/frisky/icebreaker/social/ChatActivity.java
@@ -1,9 +1,9 @@
-package com.frisky.icebreaker.ui.profile;
+package com.frisky.icebreaker.social;
 
 import android.graphics.Typeface;
-import android.os.Bundle;
 import android.support.v4.content.res.ResourcesCompat;
 import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
 import android.util.TypedValue;
 import android.view.View;
 import android.widget.ImageButton;
@@ -11,23 +11,26 @@ import android.widget.TextView;
 
 import com.frisky.icebreaker.R;
 
-public class EditProfileActivity extends AppCompatActivity {
+public class ChatActivity extends AppCompatActivity {
 
-    TextView mToolbarText;
-    ImageButton mBackButton;
-    ImageButton mDoneButton;
+    private TextView mToolbarText;
+    private ImageButton mBackButton;
+    private ImageButton mMenuButton;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_edit_profile);
+        setContentView(R.layout.activity_chat);
 
         initUI();
     }
 
     private void initUI() {
         mToolbarText = findViewById(R.id.text_app_bar);
-        mToolbarText.setText(R.string.edit_profile);
+        if (getIntent().hasExtra("name")){
+            String name = getIntent().getStringExtra("name");
+            mToolbarText.setText(name);
+        }
         mToolbarText.setTextSize(TypedValue.COMPLEX_UNIT_SP, 20);
         Typeface typeface = ResourcesCompat.getFont(getApplicationContext(), R.font.museosans700);
         mToolbarText.setTypeface(typeface);
@@ -37,16 +40,11 @@ public class EditProfileActivity extends AppCompatActivity {
         mBackButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                EditProfileActivity.super.onBackPressed();
+                ChatActivity.super.onBackPressed();
             }
         });
 
-        getSupportFragmentManager()
-                .beginTransaction()
-                .replace(R.id.frame_image_grid, new EditImagesFragment())
-                .commit();
-
-        mDoneButton = findViewById(R.id.button_app_bar_right);
-        mDoneButton.setImageResource(R.drawable.round_done_24);
+        mMenuButton = findViewById(R.id.button_app_bar_right);
+        mMenuButton.setImageResource(R.drawable.round_more_vert_24);
     }
 }
