+++ app/src/main/java/com/frisky/icebreaker/social/SocialFragment.java
@@ -1,4 +1,4 @@
-package com.frisky.icebreaker.ui.social;
+package com.frisky.icebreaker.social;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
