+++ app/src/main/res/layout/activity_chat.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.social.ChatActivity">
+    tools:context=".social.ChatActivity">
 
     <include
         layout="@layout/app_bar"
