+++ app/src/main/res/layout/activity_edit_profile.xml
@@ -5,7 +5,7 @@
     android:id="@+id/coordinatorLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.profile.EditProfileActivity">
+    tools:context=".profile.EditProfileActivity">
 
     <include
         android:id="@+id/app_bar_edit_profile"
