+++ app/src/main/res/layout/activity_menu.xml
@@ -5,7 +5,7 @@
     android:id="@+id/order_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.orders.MenuActivity">
+    tools:context=".orders.MenuActivity">
 
     <TextView
         android:id="@+id/text_table"
