+++ app/src/main/res/layout/activity_order.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.orders.OrderActivity">
+    tools:context=".orders.OrderActivity">
 
     <ImageButton
         android:id="@+id/button_back"
