+++ app/src/main/res/layout/activity_pub.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".ui.pubs.PubActivity">
+    tools:context=".pubs.PubActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar_container"
