+++ app/src/main/res/layout/activity_setup_profile.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.profile.SetupProfileActivity">
+    tools:context=".profile.SetupProfileActivity">
 
     <ImageButton
         android:id="@+id/button_cancel"
