+++ app/src/main/res/layout/content_pub.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".ui.pubs.PubActivity"
+    tools:context=".pubs.PubActivity"
     tools:showIn="@layout/activity_pub">
